requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.7.1
python-multipart==0.0.19
email-validator==2.2.0

ucimlrepo==0.0.7
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
pillow==11.1.0

torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
